# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Browser4 WebDriver-Compatible API
  description: |
    A WebDriver-compatible HTTP API with selector-first extensions for browser automation.
    This API provides session management, navigation, element interaction, and event handling
    capabilities.
  version: 1.0.0
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: http://localhost:8182
    description: Local development server
...
paths:
...
  /session/{sessionId}/url:
    post:
      tags:
        - navigation
      summary: Navigate to a URL
      operationId: navigate
      parameters:
        - $ref: '#/components/parameters/sessionId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetUrlRequest'
      responses:
        '200':
          description: Navigation successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebDriverResponse'
        '404':
          description: Session not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    get:
      tags:
        - navigation
      summary: Get current URL
      operationId: getCurrentUrl
      parameters:
        - $ref: '#/components/parameters/sessionId'
      responses:
        '200':
          description: Current URL
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UrlResponse'
        '404':
          description: Session not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /session/{sessionId}/bringToFront:
    post:
      tags:
        - navigation
      summary: Bring window to front
      operationId: bringToFront
      parameters:
        - $ref: '#/components/parameters/sessionId'
      responses:
        '200':
          description: Window brought to front
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebDriverResponse'
        '404':
          description: Session not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /session/{sessionId}/selectors/waitFor:
    post:
      tags:
        - selectors
      summary: Wait for selector to appear
      operationId: waitForSelector
      parameters:
        - $ref: '#/components/parameters/sessionId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WaitForRequest'
      responses:
        '200':
          description: Selector found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebDriverResponse'
        '408':
          description: Timeout waiting for selector
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /session/{sessionId}/capture:
    post:
      tags:
        - pulsar
      summary: Capture live page snapshot
      description: |
        Captures the live DOM from the current WebDriver session and returns a
        WebPageResult-like payload, optionally including HTML, text, or screenshot.
      operationId: capture
      parameters:
        - $ref: '#/components/parameters/sessionId'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CaptureRequest'
      responses:
        '200':
          description: Snapshot captured
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CaptureResponse'
        '404':
          description: Session not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  parameters:
    sessionId:
      name: sessionId
      in: path
      required: true
      schema:
        type: string
      description: Session identifier
...
  schemas:
...
    WebDriverResponse:
      type: object
      properties:
        value:
          description: Response value (can be null for success)
    ErrorResponse:
      type: object
      properties:
        value:
          type: object
          properties:
            error:
              type: string
            message:
              type: string
            stacktrace:
              type: string
    SetUrlRequest:
      type: object
      required:
        - url
      properties:
        url:
          type: string
          format: uri
          description: URL to navigate to
    UrlResponse:
      type: object
      properties:
        value:
          type: string
          description: Current URL
...
    WaitForRequest:
      type: object
      required:
        - selector
      properties:
        selector:
          type: string
        strategy:
          type: string
          enum: [css, xpath]
          default: css
        timeout:
          type: integer
          default: 30000
          description: Timeout in milliseconds
...
    CaptureRequest:
      type: object
      properties:
        url:
          type: string
          description: Optional URL label for the capture result
        args:
          type: string
          description: Optional load arguments (reserved for future use)
        includeHtml:
          type: boolean
          default: true
          description: Whether to include raw HTML in the response
        includeText:
          type: boolean
          default: false
          description: Whether to include extracted text in the response
        includeScreenshot:
          type: boolean
          default: false
          description: Whether to include a screenshot in the response
        screenshotSelector:
          type: string
          default: body
          description: CSS selector to scope screenshot capture
        domSettleTimeoutMs:
          type: integer
          format: int64
          description: Timeout in ms for DOM to settle before capture
        timeoutMs:
          type: integer
          format: int64
          description: Overall timeout in ms for capture
    CaptureResponse:
      type: object
      properties:
        value:
          allOf:
            - $ref: '#/components/schemas/WebPageResult'
            - type: object
              properties:
                html:
                  type: string
                  description: Raw HTML content (optional)
                text:
                  type: string
                  description: Extracted text content (optional)
                screenshot:
                  type: string
                  format: byte
                  description: Base64-encoded PNG screenshot (optional)
...